회원가입, 로그인 오류 수정

+++ src/main/java/com/uriel/travel/service/ProductService.java
@@ -205,21 +205,23 @@ public class ProductService {
         productDetail.update(requestDto);
     }
 
-
     // 상품 일괄 삭제
     public void delete(List<Long> ids) {
         ids.forEach(productId -> {
-            Product product = getProductByProductId(productId);
+            List<Order> orderList = orderRepository.findByProductId(productId);
+            if (!orderList.isEmpty()) {
 
-            Package aPackage = packageRepository.findById(product.getAPackage().getId())
-                    .orElseThrow(() ->
-                            new CustomNotFoundException(ErrorCode.NOT_FOUND));
+                Product product = getProductByProductId(productId);
+                Package aPackage = packageRepository.findById(product.getAPackage().getId())
+                        .orElseThrow(() ->
+                                new CustomNotFoundException(ErrorCode.NOT_FOUND));
 
-            ProductDetail productDetail = productDetailRepository.findByProductId(productId);
-            productDetailRepository.delete(productDetail);
+                ProductDetail productDetail = productDetailRepository.findByProductId(productId);
+                productDetailRepository.delete(productDetail);
 
-            aPackage.getProductList().remove(product);
-            productRepository.delete(product);
+                aPackage.getProductList().remove(product);
+                productRepository.delete(product);
+            }
         });
     }
